+++ packages/editor/src/components/post-trash/panel.js
@@ -1,0 +1,13 @@
+/**
+ * Internal dependencies
+ */
+import PostTrashCheck from './check';
+import PostTrashLink from './';
+
+export default function PostTrashPanel() {
+	return (
+		<PostTrashCheck>
+			<PostTrashLink />
+		</PostTrashCheck>
+	);
+}