+++ packages/commands/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Internal
+
+-   Replaced `classnames` package with the faster and smaller `clsx` package ([#61138](https://github.com/WordPress/gutenberg/pull/61138)).
+
 ## 0.28.0 (2024-05-02)
 
 ## 0.27.0 (2024-04-19)